Migrate to poetry (+ conda + pip requirements)
This is to use the good potential of
poetry in managing packaging (including
poetry.lock
for reproducbility).

So how this works now is that there is
a Makefile with install and uninstall commands:

- make install sets a new conda env,
installs poetry there, and then adds
all packages from requirements.txt
and installs them through poetry.
- make uninstall removes entire conda env.

This set-up makes use of poetry but
keeps the local dir nice and clean.

The templates for CONDA and POETRY
are taken from the new templates dir
within package, which will now also
be reused by the package.

Also, requirements.txt is now lean
and only includes packages that
are actually used by the CLI tool, and
all the LLM/data related packages
will remain within REQUIREMENTS template (to be added).