[MAJOR] change response dump format to json
This leads to the LLM response being saved
(ultimately in the YAML file) as a simple
JSON object, not a Pythonic object.

It is thus not fully possible to load a Pythonic var
after this change - it is, however, simpler to
load the YAML for other purposes such as evals,
and it looks cleaner because class definitions
are absent.

This change is not visible with simple schemas
such as the one from example 1.