[MAJOR] Allow nested evals and change manual eval logic
Previously this would throw an exception. Yet nested
evals are very helpful to replicate the structure
of the response schema in manual evals.

In terms of the new manual eval logic, now this
expects a pass and a score above zero, otherwise
it will put a fail (unless it is a skip), and this now
returns a clear label to the user that shows
the expectation and the score.

Also, it is now possible to either define a dict with
explicit manual_result, manual_score, manual_reason
or stick to a one-liner pass or fail (score will be set
to 1.0 if pass and to 0.0 if fail if not set).