MAJOR refactor of llm_provider: add openrouter etc.
Adding this as a single chunk
because was debugging in one of initialized
gerpa repos.

What this does, in major parts:

- openrouter now works
- pdf is still supported if it is
by provider, but if not image is OCR'd
and passed together with text layer;
this works for all provider because
shared logic is now a part of
BaseLLMProvider
- note that filenames of pdf/images
are/may be passed to LLM
- clean json content is now a function
that smartly parses content from
providers not supporting constrained
decoding (basically google ai studio,
but also can optionally skip constrained
decoding now if want for any provider)
- consistent (hopefully) saving of
raw request and response for all
providers, as json strings
- where raw request is really long, it is
also redacted for length now
- for ollama, details go into provider
metadata or model metadata now
as applicable, not a provider-specific
item, which reinforces common schema
- some log clarifications, and logger
is now used by functions throughout
because setup logger is global now
(and is also exported for import btw)
- prompt hashing function now returns
'reused' boolean also to easily track
if existing prompt file was used
(useful for analyses)
- multiple bug fixes throughout, including
some really bad issues with dumps, e.g.,
inability to json serialize custom classes
or badly formatted multiline yaml

This probably requires an updated experiment
notebook template, which I will add shortly.
But perhaps older syntax would also work -
did not test and not sure.